+++ src/Authorization/PolicyContext.php
@@ -4,19 +4,15 @@ namespace Orisai\Auth\Authorization;
 
 use Orisai\Auth\Authentication\Data\ExpiredLogin;
 
+/**
+ * @readonly
+ */
 interface PolicyContext
 {
 
 	public function isCurrentUser(): bool;
 
 	public function getAuthorizer(): Authorizer;
-
-	public function addAccessEntry(AccessEntry $entry): void;
-
-	/**
-	 * @return list<AccessEntry>
-	 */
-	public function getAccessEntries(): array;
 
 	/**
 	 * @return array<ExpiredLogin>