+++ tests/Unit/Authentication/BaseFirewallTest.php
@@ -11,6 +11,7 @@ use Orisai\Auth\Authentication\IntIdentity;
 use Orisai\Auth\Authentication\LogoutCode;
 use Orisai\Auth\Authentication\StringIdentity;
 use Orisai\Auth\Authorization\AccessEntry;
+use Orisai\Auth\Authorization\AccessEntryType;
 use Orisai\Auth\Authorization\AuthorizationDataBuilder;
 use Orisai\Auth\Authorization\PolicyManager;
 use Orisai\Auth\Authorization\PrivilegeAuthorizer;
@@ -147,14 +148,30 @@ final class BaseFirewallTest extends TestCase
 
 		$firewall->login(new IntIdentity(1, []));
 
-		$firewall->isAllowed(NoRequirementsPolicy::getPrivilege(), null, $entries);
-		self::assertSame([], $entries);
-
-		$firewall->isAllowed(AddAccessEntriesPolicy::getPrivilege(), null, $entries);
+		$allowed = $firewall->isAllowed(NoRequirementsPolicy::getPrivilege(), null, $entries);
+		self::assertTrue($allowed);
 		self::assertEquals(
 			[
-				new AccessEntry('Message'),
-				new AccessEntry(new TranslatableMessage('message.id')),
+				new AccessEntry(
+					AccessEntryType::allowed(),
+					'',
+				),
+			],
+			$entries,
+		);
+
+		$allowed = $firewall->isAllowed(AddAccessEntriesPolicy::getPrivilege(), null, $entries);
+		self::assertTrue($allowed);
+		self::assertEquals(
+			[
+				new AccessEntry(
+					AccessEntryType::allowed(),
+					'Message',
+				),
+				new AccessEntry(
+					AccessEntryType::allowed(),
+					new TranslatableMessage('message.id'),
+				),
 			],
 			$entries,
 		);